merge develop into main

+++ lib/sensors/sps30.cpp
@@ -63,7 +63,6 @@ void sps30::tick() {
     }
 
     if (anyChannelNeedsSampling()) {
-        // clearNewMeasurements();
         startSampling();
         state = sensorDeviceState::sampling;
     } else {
@@ -190,7 +189,7 @@ void sps30::goSleep() {
 
 bool sps30::testI2cAddress(uint8_t addressToTest) {
 #ifndef generic
-    return (HAL_OK == HAL_I2C_IsDeviceReady(&hi2c2, addressToTest << 1, halTrials, halTimeout));
+    return (HAL_OK == HAL_I2C_IsDeviceReady(&hi2c2, static_cast<uint16_t>(addressToTest << 1), halTrials, halTimeout));
 #else
     return true;
 #endif
